appease harmless compiler warnings
- sine warning from newer Clang
- LRESULT size mismatch from 64-bit Windows

The code passes a subset of `-Wextra' on GCC 8.8.1.

+++ src/baseq2/p_view.c
@@ -960,7 +960,7 @@ void ClientEndServerFrame(edict_t *ent)
         bobtime *= 4;
 
     bobcycle = (int)bobtime;
-    bobfracsin = fabsf(sin(bobtime * M_PI));
+    bobfracsin = fabsf(sinf(bobtime * M_PI));
 
     // detect hitting the floor
     P_FallingDamage(ent);
